requirements.txt
amqp==5.0.1
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
billiard==3.6.3.0
CacheControl==0.12.6
cachy==0.3.0
celery==5.0.1
certifi==2020.6.20
chardet==3.0.4
cleo==0.8.1
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
clikit==0.6.2
crashtest==0.3.1
decorator==4.4.2
distlib==0.3.1
Django==3.1.2
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
dnspython==1.16.0
eventlet==0.29.1
filelock==3.0.12
greenlet==0.4.17
gunicorn==20.0.4
html5lib==1.1
idna==2.10
importlib-metadata==1.7.0
install==1.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
keyring==21.4.0
kombu==5.0.2
lockfile==0.12.2
msgpack==1.0.0
packaging==20.4
parso==0.7.1
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.6.0
poetry-core==1.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.7.2
PyJWT==1.7.1
pylev==1.3.0
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
shellingham==1.3.2
six==1.15.0
sqlparse==0.4.1
tomlkit==0.7.0
traitlets==5.0.5
urllib3==1.25.11
vine==5.0.0
virtualenv==20.0.35
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.3.2
